:heavy_plus_sign: Reconfigure CodeBlocks
The `/projet/projet.cbp` file was moved to `/project.cbp`, was both manually edited and completely set up again with CodeBlocks, in order to (hopefully) properly include and link against SDL2 & co.

SDL2_gfx is not included. I am planning to not use it, partially due to how unprofessional it looks and because it is impossible to compile with the tools I want to use.

The "Release" build type will statically link the libraries and warn about dead code.